db/connection.py
import psycopg2
from dotenv import load_dotenv
import os

load_dotenv()

def connect_db():
    connect = psycopg2.connect(
        dbname=os.environ.get('DB_NAME'),
        user=os.environ.get('DB_USER'),
        password=os.environ.get('DB_PASSWORD'),
        host=os.environ.get('DB_HOST'),
        port=os.environ.get('DB_PORT')
    )
    return connect